amazon-sephora-pipeline/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
textblob>=0.17.1
vaderSentiment>=3.3.2
scikit-learn>=1.2.0
seaborn>=0.12.0
matplotlib>=3.6.0
plotly>=5.13.0
streamlit>=1.22.0
pyyaml>=6.0
scipy>=1.10.0